Move figure functions to new file
